Add files via upload
The Unified Compendium is a repository of > 3800 experimentally acquired CCS values obtained from traceable molecular standards and measured with drift tube-mass spectrometers. Represented in the Compendium are 14 structurally-based chemical super classes, consisting of a total of 80 classes and 157 subclasses. Using this large data set, regression fitting and predictive statistics have been performed to describe mass-CCS correlations specific to each chemical ontology. These structural trends provide a rapid and effective filtering method in the traditional untargeted workflow for identification of unknown biochemical species. The predictive abilities of this Compendium will improve in specificity and expand across more chemical classes as data from the IM-MS community is contributed. Inclusion criteria and instructions for data submission to the Compendium can be found in the 'Data Submission Guidelines and Tools' page on the menu to the left.

Please use the following for citation purposes:
J. A. Picache, B. S. Rose, A. Balinski, K. L. Leaptrot, S. D. Sherrod, J. C. May and J. A. McLean, Chem. Sci., 2019, 10, 983–993.